% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_capture_rates.R
\name{calc_unweighted_capture_rates_within_timeframes}
\alias{calc_unweighted_capture_rates_within_timeframes}
\title{Function to calculate the unweighted capture rate within specific timeframes in a year. 
Function works similar to estimate_effort}
\usage{
calc_unweighted_capture_rates_within_timeframes(data, timeframes, years = NULL)
}
\arguments{
\item{data}{Input data: Dataframe with the following collumns expected: ...}

\item{timeframes}{list of timeframes to analyze. A timeframe consists of a (start,end)-tupel in the format ('%m-%d', '%m-%d'). Timeframes should not overlap.}

\item{years}{list of years to analyze within the data. If no years are specified, all occurring years in the data are taken}
}
\value{
list of results for each year and timeframe
}
\description{
Unweighted Capture Rate: Number of observations per species in a given location divided by the number of observations
}
