% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_capture_rates.R
\name{calc_weighted_capture_rates_based_on_effort}
\alias{calc_weighted_capture_rates_based_on_effort}
\title{Calculate the weighted capture rates based on the data created by estimate_effort_within_timeframes}
\usage{
calc_weighted_capture_rates_based_on_effort(effort_data)
}
\arguments{
\item{effort_data}{expected columns: deploymentID, scientificName, effort, locationName}
}
\value{
dataframe with cr_weighted collumn containing the weighted capture rate
}
\description{
Calculate the weighted capture rates based on the data created by estimate_effort_within_timeframes
}
