% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_effort_within_timeframes.R
\name{estimate_effort_within_timeframes}
\alias{estimate_effort_within_timeframes}
\title{Function to estimate the effort within specific timeframes in a year.
If a timeframe covers two years, important is the start of the timeframe.
(e.g timeframe: ('12-01', '01-15'), years: (2015, 2018, ...) -> timeframes to analyze: (2015-12-01, 2016-01-15), (2018-12-01, 2019-01-15) ...)}
\usage{
estimate_effort_within_timeframes(data, timeframes, years = NULL)
}
\arguments{
\item{data}{Input data: Dataframe with the following collumns expected: StartDate, EndDate, SequDate, Detection_Distance.}

\item{timeframes}{list of timeframes to analyze. A timeframe consists of a (start,end)-tupel in the format ('%m-%d', '%m-%d'). Timeframes should not overlap.}

\item{years}{list of years to analyze within the data. If no years are specified, all occurring years in the data are taken}
}
\value{
list of results for each year and timeframe
}
\description{
If the Start / End Date of a deployment lie outside the timeframe boundaries, the start / end will be adjusted to be the first / last day of the period. 
Sequences will be filtered as well (based on the startDate of the sequence) and the Timespan of a deployment will be estimated based on the adjusted dates. 
Deployments that lie entierly outside of the given period will be disregarded
}
\details{
Effort = detection distance (m) * timespan (days)
}
\examples{
\dontrun{
# define test_data
test_data <- data.frame(
  deploymentID = c("06ee9610-93c3-4de5-ade7-0a507b94ad71","06ee9610-93c3-4de5-ade7-0a507b94ad71","044c34ac-7082-4240-9457-ac865aa55fcc", "044c34ac-7082-4240-9457-ac865aa55fcc", "0009d0dc-6337-47d8-8bb4-10da7bb786b4"),
  StartDate = as.Date(c("2015-01-15", "2015-03-05", "2016-03-10", "2019-08-20", "2020-02-05"), format = "\%Y-\%m-\%d"),
  EndDate = as.Date(c("2015-02-28", "2015-05-07","2016-11-30", "2019-12-31", "2020-12-31"), format = "\%Y-\%m-\%d"),
  SequDate = c(as.POSIXct(strptime("2015-01-17 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S")), as.POSIXct(strptime("2015-03-05 06:30:00", "\%Y-\%m-\%d \%H:\%M:\%S")), as.POSIXct(strptime("2016-04-12 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S")),as.POSIXct(strptime("2019-08-20 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S")),as.POSIXct(strptime("2020-02-27 01:30:00", "\%Y-\%m-\%d \%H:\%M:\%S"))),
  Detection_Distance = c(10, 5, 15, 10, 20)
  )
# define timeframes
timeframes <- list(c('02-01', '04-30'), c('05-01', '08-31'), c('09-01', '12-31'))

# (optional) define years
years <- c(2016,2017,2018)

results <- estimate_effort_within_timeframe(data, years, timeframes)
}
}
