% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_data_to_timeframes.R
\name{truncate_data_to_timeframes}
\alias{truncate_data_to_timeframes}
\title{Function to truncate data to specific timeframes in a year.
If a timeframe covers two years, important is the start of the timeframe.
(e.g timeframe: ('12-01', '01-15'), years: (2015, 2018, ...) -> timeframes to analyze: (2015-12-01, 2016-01-15), (2018-12-01, 2019-01-15) ...)}
\usage{
truncate_data_to_timeframes(data, timeframes, years = NULL)
}
\arguments{
\item{data}{Input data: Dataframe with the following collumns expected: StartDate, EndDate, SequDate}

\item{timeframes}{list of timeframes to analyze. A timeframe consists of a (start,end)-tupel in the format ('%m-%d', '%m-%d'). Timeframes should not overlap.}

\item{years}{list of years to analyze within the data. If no years are specified, all occurring years in the data are taken}
}
\value{
list of data for each year (and each timeframe per year)
}
\description{
If the Start / End Date of a deployment lie outside the timeframe boundaries, the start / end will be adjusted to be the first / last day of the period. 
Sequences will be filtered as well (based on the startDate of the sequence) and the Timespan of a deployment will be estimated based on the adjusted dates. 
Deployments that lie entierly outside of the given period will be disregarded
}
